+++ src/utils/transpose.cc
@@ -17,22 +17,24 @@
 */
 
 #include "utils/transpose.h"
+#include <cstddef>
 
 namespace lczero {
 void TransposeTensor(const std::vector<int>& dims, std::vector<int> order,
                      const std::vector<float> from, float* to) {
   if (order.empty()) {
-    for (int i = 0; i < dims.size(); ++i) order.push_back(dims.size() - i - 1);
+    for (size_t i = 0; i < dims.size(); ++i)
+      order.push_back(dims.size() - i - 1);
   }
   std::vector<int> cur_idx(dims.size());
-  for (int _ = 0; _ < from.size(); ++_) {
+  for (size_t _ = 0; _ < from.size(); ++_) {
     size_t from_idx = 0;
     for (int i : order) {
       from_idx *= dims[i];
       from_idx += cur_idx[i];
     }
     *to++ = from[from_idx];
-    for (int i = dims.size() - 1; i >= 0; --i) {
+    for (size_t i = dims.size() - 1; i >= 0; --i) {
       if (++cur_idx[i] == dims[i]) {
         cur_idx[i] = 0;
       } else {
